+++ CHANGELOG.md
@@ -18,6 +18,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## 📚 Documentation -->
 
+# [Unreleased]
+
+## 🐛 Fixes
+- **Composition failures on environment variables in supergraph config - @pubmodmatt PR #XXXX**
+
 # [0.31.1] - 2025-05-21
 
 ## 🐛 Fixes